LottoBuyPrice 원시값 포장
